Take ShiftSelection out of DocumentSelector.ts
